requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
cryptography
passlib
PyJWT
jinja2
python-multipart
pydantic-settings
httpx==0.27.2